Refactor: use proper sizes for rgba color parameters.

+++ src/sanguinecomponents.hpp
@@ -292,7 +292,7 @@ void drawRectHalo(const Widget::DrawArgs& args, const Vec boxSize, const NVGcolo
 inline void fillSvgSolidColor(NSVGimage* svgImage, const unsigned int fillColor);
 
 // utils
-inline unsigned int rgbColorToInt(const unsigned int red, const unsigned int green, const unsigned int blue, const unsigned int alpha = 255) {
+inline unsigned int rgbColorToInt(const uint8_t red, const uint8_t green, const uint8_t blue, const uint8_t alpha = 255) {
 	return (alpha << 24) + (blue << 16) + (green << 8) + red;
 }
